backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
pydantic
google-genai
httpx
pytest
pytest-asyncio
pytest-cov
fastapi-cors
email-validator
python-multipart
